third_party/WebKit/LayoutTests/webaudio/offlineaudiocontext-suspend-resume-promise.html

<!doctype html>
<html>
  <head>
    <script src="../resources/js-test.js"></script>
    <script src="resources/compatibility.js"></script>
    <script src="resources/audio-testing.js"></script>
  </head>

  <body>
    <script>
      description('Test promise resolution of OfflineAudioContext.resume() and OfflineAudioContext.suspend().');
      window.jsTestIsAsync = true;

      var context;

      // The sample rate is multiple of the rendering quantum, so suspension
      // times fall in to the render quantum boundary.
      var renderQuantum = 128;

      var sampleRate = renderQuantum * 100;
      var renderDuration = 2;
      var scheduledSuspendTime = 0;

      // With the sample rate setting above, this ensures suspension time fall
      // in to the render quantum boundary.
      var suspendInterval = 0.25;

      context = new OfflineAudioContext(1, sampleRate * renderDuration, sampleRate);

      function onSuspended() {
        Should('context.currentTime', context.currentTime)
            .beEqualTo(scheduledSuspendTime);

        scheduledSuspendTime = context.currentTime + suspendInterval;

        // When |scheduledSuspendTime >= renderDuration|, suspending the context
        // will reject the promise.
        if (scheduledSuspendTime < renderDuration) {
          context.suspend(scheduledSuspendTime).then(onSuspended);
        } else {
          Should('context.suspend(scheduledSuspendTime)',
            context.suspend(scheduledSuspendTime)).beRejected();
        }

        context.resume();
      }

      // Initiate the suspension loop.
      context.suspend(scheduledSuspendTime).then(onSuspended);
      context.startRendering().then(function () {
        Should('context.state', context.state).beEqualTo('closed');
      }).then(finishJSTest);

      successfullyParsed = true;
    </script>

  </body>
</html>
