third_party/WebKit/LayoutTests/webaudio/offlineaudiocontext-suspend-resume-promise-expected.txt

Test promise resolution of OfflineAudioContext.resume() and OfflineAudioContext.suspend().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context.currentTime is equal to 0.
PASS context.currentTime is equal to 0.25.
PASS context.currentTime is equal to 0.5.
PASS context.currentTime is equal to 0.75.
PASS context.currentTime is equal to 1.
PASS context.currentTime is equal to 1.25.
PASS context.currentTime is equal to 1.5.
PASS context.currentTime is equal to 1.75.
PASS context.suspend(scheduledSuspendTime) rejected correctly (with InvalidStateError: cannot schedule a suspend at frame 25600 (2 seconds) because it is greater than or equal to the total render duration of 25600 frames).
PASS context.state is equal to closed.
PASS successfullyParsed is true

TEST COMPLETE
